fix macos build

+++ AWSClientRuntime/Tests/Middlewares/Sha256TreeHashMiddlewareTests.swift
@@ -14,7 +14,7 @@ class Sha256TreeHashMiddlewareTests: XCTestCase {
     func testTreeHashAllZeroes() {
         let context = HttpContextBuilder().build()
         let expectation = XCTestExpectation(description: "closure was run")
-        let bytesIn5_5MB: Int = 1024*1024*5.5
+        let bytesIn5_5MB: Int = Int(1024 * 1024 * 5.5)
         let byteArray: [UInt8] = Array(repeating: 0, count: bytesIn5_5MB)
         let byteBuffer = ByteBuffer(bytes: byteArray)
         let streamInput = StreamInput(body: ByteStream.buffer(byteBuffer))
